backend/requirements.txt
flask>=2.0.0
psutil>=5.9.0
pillow>=9.0.0
mss>=6.1.0
pynput>=1.7.6
requests>=2.27.0
psycopg>=3.1.0
cryptography>=41.0.0
